feat: comments and moving files around
