+++ packages/at_repl/lib/src/home_directory.dart
@@ -1,0 +1,23 @@
+import 'dart:io';
+
+// Get the home directory or null if unknown.
+String? getHomeDirectory() {
+  switch (Platform.operatingSystem) {
+    case 'linux':
+    case 'macos':
+      return Platform.environment['HOME'];
+    case 'windows':
+      return Platform.environment['USERPROFILE'];
+    case 'android':
+      // Probably want internal storage.
+      return '/storage/sdcard0';
+    case 'ios':
+      // iOS doesn't really have a home directory.
+      return null;
+    case 'fuchsia':
+      // I have no idea.
+      return null;
+    default:
+      return null;
+  }
+}
